Realtime Updating UI and Scene/Camera Abstraction
The UI can now update per line rendered or per pixel rendered.
The render scene and camera has been abstracted. There is still a lot
of work to be done on the scene and camera objects

Big issue: The RayTest.java file needs to be better organized. The
"dependence" on the main function may not be so good in the long run.